requirements.txt
Flask==2.2.3
python-telegram-bot==20.8
openai
beautifulsoup4
requests
